New component PlaceCategories
